Minor CSS tweak for page form.
Making the page-handle form input line up with everything else (the
preceding '/' was pushing it forward a bit).

This is really for flogiston-cms, but I didn't want to go through the
trouble of having extra CSS files and includes just for such a minor
tweak.

And really, some other recent CMS stuff has made it into here. This is
the base of all flogiston, so sometimes it gets things like this.