Namespacing controllers.
The other flogiston plugins have namespaced models because that's the
way we found to get the flogiston behavior in there with allowing the
app to add some behavior of its own (and deal with all the annoying
Rails constant loading crap).

Turns out it's a useful/important thing to do with the controllers, too.